requirements.txt
opencv-python>=4.8
numpy>=1.24
pytesseract>=0.3.10
Pillow>=10.0
matplotlib>=3.7
